+++ vkconfig_gui/CHANGELOG.md
@@ -2,7 +2,7 @@
 [Vulkan SDK 1.4.3XX.0](https://github.com/LunarG/VulkanTools/tree/main)
 
 ### Features:
-- Add dark mode support, automatically matching the OS mode
+- Add dark mode support, automatically matching the OS mode, when Qt dark mode is available on the OS
 - Add `vkconfig loader` commands to import, export, delete and use stored loader configurations
 - Add new Vulkan SDK notification and download
 
@@ -13,6 +13,8 @@
 - Add button to open launcher log file
 - Add `vulkaninfo` as default application on macOS
 - Enable text editing the executable path
+- Add layer manifest validation log saving option
+- Add ignoring invalid layer manifest when invalid until the file is modified
 
 ### Fixes:
 - Fix application arguments and environment variables splitting